refactor(tui): extract shared session rendering primitives
